% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmeans.R
\name{adjmeans}
\alias{adjmeans}
\title{Adjusted means based in BLUEs}
\usage{
adjmeans(data, trait, treatment, fix.model)
}
\arguments{
\item{data}{Experimental design data frame with the factors and traits.}

\item{trait}{Name of the trait.}

\item{fix.model}{The fixed effects in the model. See examples.}

\item{treat}{Name of the treat.}
}
\value{
A data frame with the BLUEs.
}
\description{
Function to calculate the adjusted means using the Best Linear Unbiased Estimators (BLUEs).
}
\details{
The blues calculation is based in the pairwaise comparison and its could takes time according the number of the treatments.
}
\examples{

library(agridat)
adm <- adjmeans(data = john.alpha
                , trait = "yield"
                , treatment = "gen"
                , fix.model = "rep + (1|rep:block) + gen"
                )
adm

}
\author{
Flavio Lozano-Isla
}
