% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmeans_lm.R
\name{adjmeans_lm}
\alias{adjmeans_lm}
\title{Adjusted means using linear models}
\usage{
adjmeans_lm(
  data,
  trait,
  lm.model,
  comparison,
  digits = 3,
  tab_vars = NULL,
  sep = NULL,
  plot_treat = NULL,
  plot_groups = NULL,
  plot_diag = FALSE
)
}
\arguments{
\item{data}{Experimental design data frame with the factors and traits.}

\item{trait}{Name of the trait.}

\item{lm.model}{The effects in the model. See examples.}

\item{comparison}{Factor for the comparisons.}

\item{digits}{Number of digits in the table (Default = 3).}

\item{tab_vars}{Specific the variables in the summary table (Default = NULL).}

\item{sep}{Separator between the variables when use tab_vars. See details}

\item{plot_treat}{Treat of comparison in dot plot (Default = NULL). See example.}

\item{plot_groups}{Group of comparison in dot plot (Default = NULL). See example.}

\item{plot_diag}{Show diagnostic plots (default = FALSE).}
}
\value{
A list with two objects

 1. Summary statistics

 2. Adjusted means with the mean comparisons under HSD test (p<0.05)
}
\description{
Function to calculate the adjusted means using Linear models (lm or aov).
}
\details{
Using \code{tab_vars} select the variables shown the adjusted mean table with the \code{sep}.
}
\author{
Flavio Lozano-Isla
}
