% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortext.R
\name{colortext}
\alias{colortext}
\title{Colourise text for display in the terminal.}
\usage{
colortext(text, fg = "red", bg = NULL)
}
\arguments{
\item{text}{character vector}

\item{fg}{foreground colour, defaults to white}

\item{bg}{background colour, defaults to transparent}
}
\description{
If R is not currently running in a system that supports terminal colours
the text will be returned unchanged.
}
\details{
Allowed colours are: black, blue, brown, cyan, dark gray, green, light
blue, light cyan, light gray, light green, light purple, light red,
purple, red, white, yellow
}
\examples{

print(colortext("Red", "red"))
cat(colortext("Red", "red"), "\n")
cat(colortext("White on red", "white", "red"), "\n")

}
\author{
testthat package
}
