% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_conditional_panels.R
\name{design_conditional_panels}
\alias{design_conditional_panels}
\title{Conditional Panels for statistical design}
\usage{
design_conditional_panels()
}
\description{
Differents types of statistical designs are related to customized inputs.
}
\author{
quipo.org
}
