% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{design_fieldbook}
\alias{design_fieldbook}
\title{Field book design for data collection}
\usage{
design_fieldbook(treat1 = NULL, treat2 = NULL, rep = NULL, intime = 1,
  design = "crd", lbl_treat1 = NULL, lbl_treat2 = NULL,
  variables = NULL)
}
\arguments{
\item{treat1}{strign with the name of the level factor with only space}

\item{treat2}{strign with the name of the level factor with only space}

\item{rep}{number of repetition}

\item{design}{experimental design c("rcbd","crd","lsd")}

\item{lbl_treat1}{col label for treat 1}

\item{lbl_treat2}{col label for treat 2}

\item{variables}{name of the variable evaluated in string with only space}
}
\value{
Table with the experimental design
}
\description{
Function to construct your field book for your experiment
}

