% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_split.R
\name{design_split}
\alias{design_split}
\title{Split-plot experimental designs}
\usage{
design_split(
  nfactors = 2,
  factors,
  type = "split-rcbd",
  rep = 3,
  zigzag = FALSE,
  nrows = NA,
  serie = 1000,
  seed = NULL,
  project = "inkaverse",
  qrcode = "{project}{plots}{factors}"
)
}
\arguments{
\item{nfactors}{Number of factors in the experiment.}

\item{factors}{List with factor levels.}

\item{type}{Split-plot design type.}

\item{rep}{Number of replications.}

\item{zigzag}{Field layout in zigzag.}

\item{nrows}{Experimental design dimension by rows.}

\item{serie}{Number to start the plot id.}

\item{seed}{Seed for randomization.}

\item{project}{Barcode prefix.}

\item{qrcode}{String to concatenate the QR code.}
}
\value{
A list with the fieldbook design and parameters.
}
\description{
Dispatch split-plot experimental designs.
}
