% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_file_xlsx.R
\name{export_file_xlsx}
\alias{export_file_xlsx}
\title{Export xlsx}
\usage{
export_file_xlsx(data, fname)
}
\arguments{
\item{data}{exported data}

\item{fname}{file name}
}
\description{
export fieldbook in xlsx
}
