% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_data.R
\docType{data}
\name{fb_template}
\alias{fb_template}
\title{Field book template for Fieldbook Application}
\format{A data frame data.frame which comes from importing csv-template files.}
\source{
Quipo Organization.
}
\usage{
fb_template
}
\description{
This dataset contains the template for field book generation
}
\references{
This data is related to Fieldbook application
}

