% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_report.R
\name{fielbook_report}
\alias{fielbook_report}
\title{Field book report}
\usage{
fielbook_report(
  data,
  fb_smr,
  variable,
  model_facts,
  treat_comp,
  test_comp = "SNK",
  sig_level = 0.05,
  comp_diag = NULL,
  model_diag = FALSE
)
}
\arguments{
\item{data}{Field book data.}

\item{fb_smr}{Summary of the variables in the fieldbook.}

\item{variable}{Model used for the experimental design.}

\item{model_facts}{Comparison test (default = "SNK"). Others: "TUKEY" & "DUNCAN".}

\item{treat_comp}{Significance level for the analysis (default = 0.05).}

\item{test_comp}{Significance level for the analysis (default = 0.05).}

\item{sig_level}{Significance level for the analysis (default = 0.05).}

\item{comp_diag}{Factors to compare. See details.}

\item{model_diag}{Diagnostic for model (default = FALSE).}
}
\value{
data frame
}
\description{
Function to create a complete report of the fieldbook
}
\details{
For compare the factors you should use ":". For example, to compare treatment1
and treatment2 => "treatment1:treatment2".
}
\examples{

\dontrun{

library(inti)
library(googlesheets4)
library(tidyverse)

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "15uwCgQRtR01B3FJaZBE8t_0bOC_8Bbey9ccwVlZH0jg/edit#gid=56711214")
# browseURL(url)
gs <- as_sheets_id(url)

(data <- gs \%>\%
    range_read("fb"))

(fb_smr <- gs \%>\%
  range_read("fbsmr"))

report <- fielbook_report(data
                          , fb_smr = fb_smr
                          , variable = "rendexportable"
                          , comp_diag = T
                          , model_diag = F
                          )
report

}

}
\author{
Flavio Lozano-Isla
}
