% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_mvr.R
\name{fieldbook_mvr}
\alias{fieldbook_mvr}
\title{Multivariate Analysis}
\usage{
fieldbook_mvr(data, fb_smr, summary_by, groups)
}
\arguments{
\item{data}{Field book data.}

\item{fb_smr}{Summary of the variables in the fieldbook.}

\item{summary_by}{Variables for group the analysis.}

\item{groups}{Groups for color in PCA.}
}
\value{
result and plots
}
\description{
Multivariate analysis for PCA and HCPC
}
\details{
Compute and plot information for multivariate analysis (PCA, HCPC and
correlation).
}
