% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_reshape.R
\name{fieldbook_reshape}
\alias{fieldbook_reshape}
\title{Field book reshape}
\usage{
fieldbook_reshape(
  data,
  last_factor,
  sep,
  new_colname,
  last_var = NULL,
  exc_factors = FALSE
)
}
\arguments{
\item{data}{Field book raw data.}

\item{last_factor}{the last factor in your field book.}

\item{sep}{character that separates the last value.}

\item{new_colname}{The new name for the column created.}

\item{last_var}{the last variable in case you want to exclude several variables.}

\item{exc_factors}{factor to exclude during the reshape.}
}
\value{
data frame
}
\description{
Function to reshape fieldbook according a separation character
}
\details{
If you variable name is variable_evaluation_rep.
The reshape function will help to create the column "rep"
and the new variable name will be variable_evaluation
}
\examples{

\dontrun{

library(inti)
library(googlesheets4)
library(tidyverse)

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "1gue-wSQcEu4nJigVZdUWsTfIIzhtxpDRdWAQiEHgKak/edit#gid=56106366")
# browseURL(url)
gs <- as_sheets_id(url)

(data <- gs \%>\%
    range_read("fb"))

fbrs <- fieldbook_reshape(data = data
                          , last_factor = "imbt"
                          , sep = "_"
                          , new_colname = "rep"
                          , last_var = "swc_0_1"
                          , exc_factors = "bar_code"
                          )

}

}
\author{
Flavio Lozano-Isla
}
