% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_varlist.R
\name{fieldbook_varlist}
\alias{fieldbook_varlist}
\title{Field book variable list}
\usage{
fieldbook_varlist(
  varlits,
  fblength,
  abbreviation = "abbreviation",
  evaluation = "evaluation",
  sampling = "sampling"
)
}
\arguments{
\item{varlits}{Data frame with the variables information. See examples.}

\item{fblength}{Number of row in the fieldbook  (default = NULL). See details.}

\item{abbreviation}{Column contain the variables abbreviations (default = "abbreviation"). See details.}

\item{evaluation}{Column contain when the variables will be made (default = "evaluation").}

\item{sampling}{Column contain the number of sampling in each experimental unit (default = "sampling"). See details.}
}
\value{
data frame
}
\description{
Function to include the variables to evaluate in the fieldbook design.
}
\details{
The function allows to include the arguments in the sheet that have the information of the variables.
You should include 3 columns in the sheet: "\code{{abbreviation}}", "\code{{evaluation}}" and "\code{{sampling}}". See examples.
The information will be extracted automatically and deploy the list of variable for the fieldbook design.
If you will be evaluate the experimental unit one time you should set 1.
}
\examples{

\dontrun{

library(googlesheets4)
library(tidyverse)
library(inti)

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "1ilw0NHT7mihaM-3U48KzkuMt927xe8ukX6rNuIw2fT0/edit#gid=4349630")

# browseURL(url)
gs <- as_sheets_id(url)

(varlits <- gs \%>\%
    range_read("variables"))

fb <- varlits \%>\% fieldbook_varlist(fblength = 30)

}

}
\author{
Flavio Lozano-Isla
}
