% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2qmd.R
\name{figure2qmd}
\alias{figure2qmd}
\title{Figure to Quarto format}
\usage{
figure2qmd(text, path = ".", opts = NA)
}
\arguments{
\item{text}{Markdown text with figure information [string]}

\item{path}{Image path for figures [path: "." (base directory)]}

\item{opts}{chunk options in brackets [string: NA]}
}
\value{
string mutated
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
\details{
Quarto option can be included in the title using "\{\{\}\}" separated by commas
}
