% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2rmd.R
\name{figure2rmd}
\alias{figure2rmd}
\title{Figure to Rmarkdown}
\usage{
figure2rmd(text, path = ".", opts = NA, prefix = "Figure")
}
\arguments{
\item{text}{String with the table information}

\item{path}{Path of the image fot the figure}

\item{opts}{chunk options in brackets.}

\item{prefix}{Prefix for the name of the figure}
}
\value{
Mutated string
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
