% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{footnotes}
\alias{footnotes}
\title{Table footnote}
\source{
\url{https://github.com/haozhu233/kableExtra/blob/master/R/add_footnote.R}
}
\usage{
footnotes(table, notes = NULL, label = "Note:", notation = "alphabet")
}
\arguments{
\item{table}{kable output in pandoc format.}

\item{notes}{footnotes for the table.}

\item{label}{label for start the footnote.}

\item{notation}{notation for the symbols and footnotes (default = "number"). See details.}
}
\value{
Table with footnotes for word and html documents
}
\description{
Include tables footnotes and symbols for kables in pandoc format
}
\details{
You should use the pandoc format \code{kable(format = "pipe")}.
You can add the footnote symbol using \code{{-}} in your table.
\code{notation} could be use: "alphabet", "number", "symbol", "none".
}
