% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc2qmd.R
\name{gdoc2qmd}
\alias{gdoc2qmd}
\title{Google docs to Rmarkdown}
\usage{
gdoc2qmd(file, export = NA, format = "qmd", type = "asis", fill_table = "down")
}
\arguments{
\item{file}{Zip file path from Articul8 exported in md format \verb{[path]}}

\item{export}{Path to export the files \verb{[path: NA (file directory)]}}

\item{format}{Output format \verb{[character: "qmd", "rmd"]}}

\item{type}{output file type \verb{[character: "asis" "list", "listfull", "full"]}}

\item{fill_table}{extract table in listfull \verb{[character: "up", "down"]}}
}
\value{
path
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
\details{
Document rendering until certain point: "#| end"
Include for next page: "#| newpage"
You can include the cover page params using "#|" in a Google docs table
}
