% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc2rmd.R
\name{gdoc2rmd}
\alias{gdoc2rmd}
\title{Google docs to Rmarkdown}
\usage{
gdoc2rmd(file, export = NA)
}
\arguments{
\item{file}{Zip file path from Articul8 exported in md format [path]}

\item{export}{Path to export the files [path: NA (file directory)]}
}
\value{
path
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
