% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc_rticle.R
\name{gdoc_rticle}
\alias{gdoc_rticle}
\title{Google docs to Rmarkdown}
\usage{
gdoc_rticle(file, export = "files", md = TRUE)
}
\arguments{
\item{file}{Zip file path from Articul8 exported in  md format}

\item{export}{path to export the files. Default work directory}

\item{md}{Conserve md file. Default = TRUE}
}
\value{
folder
}
\description{
Use Articul8 Add-ons from Google doct to build a Rticle
}
