% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{getData}
\alias{getData}
\title{Import google spreadsheet or xlsx file}
\usage{
getData(dir, sheet = 1)
}
\arguments{
\item{dir}{local file directory for xlsx document or url from google spreadsheet}

\item{sheet}{if is a xlsx file, you can choose the sheet number}
}
\value{
data frame
}
\description{
function to import information from google spreadsheet or xlsx file.
}
