% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huito_barcode.R
\name{huito_barcode}
\alias{huito_barcode}
\title{Barcode generator}
\usage{
huito_barcode(text, color = "black", alpha = 1)
}
\arguments{
\item{text}{text to convert to QR bar code}

\item{color}{Bar code color}

\item{alpha}{Intensity og the bar code color}
}
\value{
plot
}
\description{
Generate bar codes using QR codes
}
\examples{

\dontrun{

huito_barcode("inkaverse")

}

}
