% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huito_label_options.R
\name{huito_label_options}
\alias{huito_label_options}
\title{Label options}
\usage{
huito_label_options(layers)
}
\arguments{
\item{layers}{value to match the lists in the environment}
}
\value{
data frame
}
\description{
Get environment variables for label options
}
