% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_fig.R
\name{include_fig}
\alias{include_fig}
\title{Figure with caption and notes}
\usage{
include_fig(
  data = NULL,
  figure,
  caption = NULL,
  notes = NULL,
  label = "Source:",
  table = c("figures", "caption", "notes", "url")
)
}
\arguments{
\item{data}{Data frame with the figures information. See details.}

\item{figure}{Name or path of the figure.}

\item{caption}{Manual figure caption (default = NULL).}

\item{notes}{Manual figure notes (default = NULL).}

\item{label}{Label for start the footnote (default = "Source:").}

\item{table}{Columns in the table (default = c("figures", "caption", "notes",
"url")).}
}
\value{
Figure with caption and notes
}
\description{
Include figures with title and notes using a data base
}
\details{
The data frame should contain 4 columns specified in \code{table}:

1. \code{figures}.

2. \code{caption}.

3. \code{notes}.

4. \code{url} or \code{url}.

If you don't use a a data frame you provide the information manually.
}
