% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_figure.R
\name{include_figure}
\alias{include_figure}
\title{Figure with caption and notes}
\usage{
include_figure(
  data = NULL,
  figure,
  caption = NULL,
  notes = NULL,
  label = "Note:",
  table = c("figures", "descripton")
)
}
\arguments{
\item{data}{Data frame with the figures information. See details.}

\item{figure}{Name or path of the figure.}

\item{caption}{Manual figure caption (default = NULL).}

\item{notes}{Manual figure notes (default = NULL).}

\item{label}{Label for start the footnote (default = "Source:").}

\item{table}{Columns in the table (default = c("figure", "description")).}
}
\value{
Figure with caption and notes
}
\description{
Include figures with title and notes using a data base
}
\details{
The data frame should contain 4 rows in \code{table}:

1. \code{caption}.

2. \code{notes}.

3. \code{url} 

4. \code{path}.

If you don't use a data frame you provide the information manually.
}
\examples{

library(googlesheets4)
library(inti)

if (gs4_has_token()) {

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "1dfgpmCKdPmxRHozrZp0iE_xMGsvKTIcztDpMWYSEGaY")

gs <- as_sheets_id(url)

data <- gs \%>\% 
  range_read("fig1")

fig <- include_figure(data)
fig

}

}
