% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_pdf.R
\name{include_pdf}
\alias{include_pdf}
\title{Include PDF in markdown documents}
\usage{
include_pdf(file, width = "100\%", height = "600")
}
\arguments{
\item{file}{file path from pdf file.}

\item{width}{width preview file.}

\item{height}{height preview file.}
}
\value{
html code for markdown
}
\description{
Insert PDF files in markdown documents
}
