% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_figure.R
\name{info_figure}
\alias{info_figure}
\title{Figure information}
\usage{
info_figure(caption = NA, notes = NA, url = NA, path = NA, nrows = 50)
}
\arguments{
\item{caption}{Figure caption.}

\item{notes}{Figure notes.}

\item{url}{Figure url or link.}

\item{path}{Figure local path.}

\item{nrows}{Free rows in the export table.}
}
\value{
data frame
}
\description{
Describe figures information for export or print
}
\examples{

finfo <- info_figure(
  caption = "Figure caption"
  , url = "www.image.dir"
  , nrows = 30
  )


}
