% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_table.R
\name{info_table}
\alias{info_table}
\title{Table information}
\usage{
info_table(data, caption = NA, notes = NA)
}
\arguments{
\item{data}{data frame}

\item{caption}{Figure caption.}

\item{notes}{Figure notes.}
}
\value{
data frame
}
\description{
Add table information for include_table
}
\examples{

tab <- tibble(
x = rep_len("a", 5)
, y = rep_len("b", 5)
, z = rep_len("c", 5)
)

tinfo <- tab \%>\% 
info_table(
  caption = "Figure caption"
  , notes = "notes"
  )

}
