% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jc_tombola.R
\name{jc_tombola}
\alias{jc_tombola}
\title{Journal Club Tombola}
\usage{
jc_tombola(
  data,
  members,
  group,
  gr_lvl,
  status,
  st_lvl,
  frq,
  date,
  seed = NULL
)
}
\arguments{
\item{data}{Data frame withe members and their information}

\item{members}{Columns with the members names}

\item{group}{Column for arrange the group}

\item{gr_lvl}{Levels in the groups for the arrange}

\item{status}{Column with the status of the members}

\item{st_lvl}{Level to confirm the assistance in the JC}

\item{frq}{Number of the day for each session}

\item{date}{Date when start the first session of JC}

\item{seed}{Number for replicate the results}
}
\value{
data frame with the schedule for the JC
}
\description{
Function for arrange journal club schedule
}
\author{
Flavio Lozano-Isla
}
