% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_eqn.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Regression line equation}
\usage{
lm_eqn(x, y, data)
}
\arguments{
\item{x}{variable in the x axis}

\item{y}{variable in the y axis}

\item{data}{dataframe with the information}
}
\value{
regression equation
}
\description{
Construction of the regression line equation
}
\author{
Flavio Lozano-Isla
}
