% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenMon.R
\name{PenMon}
\alias{PenMon}
\title{Penman-Monteith method}
\usage{
PenMon(
  data,
  date,
  Tmin,
  Tmax,
  RHmin,
  RHmax,
  sunshine,
  wind,
  lat,
  alt,
  Hws,
  alb = 0.23
)
}
\arguments{
\item{data}{data frame with the metrerological information.}

\item{date}{date of the weather measure. format: yyy-mm-dd.}

\item{Tmin}{temperatura minima (ºC).}

\item{Tmax}{temperatura maxima (ºC).}

\item{RHmin}{humedad relativa minima (\%).}

\item{RHmax}{humedad relativa maxima (\%).}

\item{sunshine}{sunshine (hours).}

\item{wind}{wind velocity (m.seg-1).}

\item{lat}{latitud (deciaml radians).}

\item{alt}{altitud (meters)}

\item{Hws}{heigth of data collect for wind speed (meters).}

\item{alb}{albedo}
}
\value{
data frame with daily evapotranspiration
}
\description{
Function for calculate the daily evapotranspiration.
}
\author{
Flavio Lozano-Isla
}
