% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_brln}
\alias{plot_brln}
\title{Plot line or bar graphic}
\usage{
plot_brln(data, type = "bar", x, y, z, ylab = NULL, xlab = NULL,
  lgl = NULL, lgd = "top", sig = NULL, erb = FALSE, lmt = NULL,
  brk = NULL, xbl = NULL, zbl = NULL, color = TRUE, font = 1)
}
\arguments{
\item{data}{Output dtsm fuction}

\item{type}{Type of graphic. "bar" or "line"}

\item{x}{Axis x variable}

\item{y}{Axis y variable}

\item{z}{Group variable}

\item{ylab}{Title for the axis y}

\item{xlab}{Title for the axis x}

\item{lgl}{Title for the legend}

\item{lgd}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{sig}{Significance of the result (letters)}

\item{erb}{Show the error bar.}

\item{lmt}{limits of the y axis}

\item{brk}{break of the y axis}

\item{xbl}{axis brakes labels in strign with doble space}

\item{zbl}{legend label in strign with doble space}

\item{color}{colored figure (TRUE), otherwise black & white (FALSE)}

\item{font}{letter size in plot}
}
\value{
Line o bar plot
}
\description{
Function use the dtsm funtion for plot the results
}

