% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Correlation graphic}
\usage{
plot_correlation(data, method = "pearson", sig = 0.05, color = NULL,
  font = 1)
}
\arguments{
\item{data}{matrix with numeric data}

\item{method}{method of the correlation analisis: c("pearson", "kendall", "spearman", "lin")}

\item{sig}{level of significance}

\item{color}{colors in the correlation matrix}
}
\value{
correlation plot
}
\description{
Function use to plot correlation matrix
}

