% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diag.R
\name{plot_diag}
\alias{plot_diag}
\title{Diagnostic plots}
\usage{
plot_diag(model, title = NA)
}
\arguments{
\item{model}{Statistical model}

\item{title}{Plot title}
}
\value{
plots
}
\description{
Function to plot the diagnostic of models
}
\examples{

\dontrun{

library(inti)

lm <- aov(stemdw ~ bloque + geno*treat, data = potato)

#lm <- lme4::lmer(stemdw ~ bloque + (1|geno:treat), data = potato)
 
plot(lm, which = 1)
plot_diag(lm)[3]

plot(lm, which = 2)
plot_diag(lm)[2]

plot(lm, which = 3)
plot_diag(lm)[4]

plot(lm, which = 4)
plot_diag(lm)[1]

}

}
