% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gr_dev.R
\name{plot_gr}
\alias{plot_gr}
\title{Plot line or bar graphic}
\usage{
plot_gr(
  data,
  type = "bar",
  x,
  y,
  group,
  x_lab = NULL,
  y_lab = NULL,
  g_lab = NULL,
  lgd = "top",
  sig = NULL,
  erb = FALSE,
  y_lmt = NULL,
  y_brk = NULL,
  x_brk = NULL,
  g_brk = NULL,
  color = TRUE,
  ang = 0,
  font = 1.5
)
}
\arguments{
\item{data}{Output dtsm fuction}

\item{type}{Type of graphic. "bar" or "line"}

\item{x}{Axis x variable}

\item{y}{Axis y variable}

\item{group}{Group variable}

\item{x_lab}{Title for the axis x}

\item{y_lab}{Title for the axis y}

\item{g_lab}{Title for the legend}

\item{lgd}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{sig}{Significance of the result (letters)}

\item{erb}{Show the error bar.}

\item{y_lmt}{limits of the y axis}

\item{y_brk}{break of the y axis}

\item{x_brk}{axis brakes labels in strign with doble space}

\item{g_brk}{legend label in strign with doble space}

\item{color}{colored figure (TRUE), otherwise black & white (FALSE)}

\item{ang}{x text label angle}

\item{font}{letter size in plot}
}
\value{
Line o bar plot
}
\description{
Function for present the results in line or bar plot
}
