% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_linereg}
\alias{plot_linereg}
\title{Plot line regression}
\usage{
plot_linereg(data, x, y, z, ylab = NULL, xlab = NULL, lgl = NULL,
  lgd = "top", xbrk = NULL, ybrk = NULL, zbl = NULL, color = TRUE,
  font = 1, rlx = NULL, rly = NULL)
}
\arguments{
\item{data}{Output dtsm fuction}

\item{x}{Axis x variable}

\item{y}{Axis y variable}

\item{z}{Group variable}

\item{ylab}{Title for the axis y}

\item{xlab}{Title for the axis x}

\item{lgl}{Title for the legend}

\item{lgd}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{xbrk}{brakes for x axis}

\item{ybrk}{brakes for y axis}

\item{zbl}{legend label in strign with doble space}

\item{color}{colored figure (TRUE), otherwise black & white (FALSE)}

\item{font}{letter size in plot}

\item{rlx}{regression line position in axis x.}

\item{rly}{regression line position in axis y.}
}
\value{
Line regression plot
}
\description{
Function plot linea regression
}
