% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Principal component analisys graphic}
\usage{
plot_PCA(data, type = "biplot", quali.sup = NULL, lgl = NULL)
}
\arguments{
\item{data}{matrix with numeric data}

\item{type}{type of plot PCA: c("ind", "var", "biplot")}

\item{quali.sup}{number of colum of qualitative variable}

\item{lgl}{label legend}
}
\value{
PCA biplot graph
}
\description{
Function use to plot biplot principal component analisys
}

