% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_outliers.R
\name{rm_outliers}
\alias{rm_outliers}
\title{Remove outliers}
\usage{
rm_outliers(data, trait, n_low, n_high)
}
\arguments{
\item{data}{Experimental design data frame with the factors and traits.}

\item{trait}{Name of the trait.}

\item{n_low}{Number of point to remove with lowest values}

\item{n_high}{Number of point to remove with highest values}
}
\value{
Vector with the selected values as NA's
}
\description{
Function to remove outliers according the lowest or highest values
}
\details{
Could be use applying pipes. Remove n lowest or highest point.
}
\author{
Flavio Lozano-Isla
}
