% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rticles_addin.R, R/tarpuy.R
\name{rticles_addin}
\alias{rticles_addin}
\title{Build markdown o bookdown template}
\usage{
rticles_addin()

rticles_addin()
}
\value{
Shiny app

Shiny app
}
\description{
Invoke RStudio add-in to create markdown documents

Invoke RStudio add-in to create field book designs
}
\details{
Create all the files in the present work directory.
It is recommended use ".Rproj" for bookdown documents.
After create the ".Rproj" file.
Open the ".Rprj" and you should have the possibility to compile the book.

Tarpuy allow to create experimental designs under an interactive app.
}
\examples{
\dontrun{

 inti::rticles()

}

\dontrun{

 inti::tarpuy()

}

}
\author{
Flavio Lozano-Isla

Flavio Lozano-Isla
}
