% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rticles_files.R
\name{rticles_files}
\alias{rticles_files}
\title{Files paths for shiny app}
\usage{
rticles_files(name, project = TRUE, type = c("markdown", "bookdown"))
}
\arguments{
\item{name}{file name}

\item{project}{create project}

\item{type}{type of document}
}
\value{
file list
}
\description{
List of files for shiny app
}
