% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sm.R, R/utils.R
\name{stat_sm}
\alias{stat_sm}
\title{Descriptive Statistics for a model}
\usage{
stat_sm(modelo, data)

stat_sm(modelo, data)
}
\arguments{
\item{modelo}{an object containing the results returned by a model fitting function}

\item{data}{data set used for the model}
}
\value{
data frame

data frame
}
\description{
Function to summary descriptive statistics from a model

Function to summary descriptive statistics from a model
}
