% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2qmd.R
\name{table2qmd}
\alias{table2qmd}
\title{Table to Quarto format}
\usage{
table2qmd(text, type = "asis")
}
\arguments{
\item{text}{Markdown text with table information (string)}

\item{type}{output file type [strig: "asis" "list", "listfull", "full"]}
}
\value{
string mutated
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
