% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2rmd.R
\name{table2rmd}
\alias{table2rmd}
\title{Table to Rmarkdown format}
\usage{
table2rmd(text, opts = NA)
}
\arguments{
\item{text}{String with the table information}

\item{opts}{chunk options in brackets.}
}
\value{
Mutated string
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
