% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarpuy.R
\name{tarpuy}
\alias{tarpuy}
\title{Interactive fieldbook designs}
\usage{
tarpuy(dependencies = FALSE)
}
\arguments{
\item{dependencies}{Logical. If \code{TRUE}, install any missing TARPUY runtime
dependencies using the local \code{inst/tarpuy/setup.R} script included with
the installed \code{inti} package.}
}
\value{
A Shiny application launched with \code{\link[shiny:runApp]{shiny::runApp()}}.
}
\description{
Invoke RStudio addin to create fieldbook designs
}
\details{
TARPUY allows users to create experimental designs through an interactive
Shiny application.
}
\examples{

if (interactive()) {
  inti::tarpuy()
}

}
