% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarpuy_fbapp.R
\name{tarpuy_fbapp}
\alias{tarpuy_fbapp}
\title{Field book app connection}
\usage{
tarpuy_fbapp(fieldbook = NULL, last_factor = NULL, traits = NULL)
}
\arguments{
\item{fieldbook}{Experiment field book [dataframe].}

\item{last_factor}{Last factor in the field book [string: colnames]}

\item{traits}{Traits information [dataframe or list].}
}
\value{
list
}
\description{
Function to export field book and traits for be used in field book app.
}
\details{
For the traits parameters you can used shown in the field Book app
}
\examples{

library(inti)

fieldbook <- inti::potato

traits <- list(
  list(variable = "altura de planta"
       , abbreviation = "altp"
       , format = "numeric"
       , when = "30, 40, 50"
       , samples = 3
       , units = "cm"
       , details = NA
       , minimum = 0
       , maximum = 100
       )
  , list(variable = "severidad"
         , abbreviation = "svr"
         , format = "categorical"
         , when = "30, 40, 50"
         , samples = 1
         , units = "scale"
         , details = NA
         , categories = "1, 3, 5, 7, 9"
  )
  ,  list(variable = "foto"
          , abbreviation = "foto"
          , format = "photo"
          , when = "hrv, pshrv"
          , samples = 1
          , units = "image"
          , details = NA
  )
  ,  list(variable = "germinacion"
          , abbreviation = "ger"
          , format = "boolean"
          , when = "30, 40, 50"
          , samples = 1
          , units = "logical"
          , details = NA
  )
) 

fbapp <- tarpuy_fbapp(fieldbook, last_factor = "bloque", traits)

\dontrun{ 

library(inti)
library(gsheet)

url <- paste0("https://docs.google.com/spreadsheets/d/"
       , "13DVdFht6sp4X5MSe7-4S8Mf-qCJFaEOOU0KbLzc0HjE/edit#gid=550178001")
       
traits <- gsheet2tbl(url) 

}

}
