% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_comparison}
\alias{test_comparison}
\title{Multiple comparison test}
\usage{
test_comparison(aov, comp, type = "snk", sig = 0.05)
}
\arguments{
\item{aov}{lm o aov result function.}

\item{comp}{treatments will be compared.}

\item{type}{method for made comparision analysis: c("snk", "tukey", "duncan").}

\item{sig}{significance level. Default 0.05}
}
\value{
Table with complete data for graphics
}
\description{
Function analisis of variance for summary data.
}

