% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_table.R
\name{web_table}
\alias{web_table}
\title{HTML tables for markdown documents}
\usage{
web_table(
  data,
  caption = NULL,
  digits = 2,
  rnames = FALSE,
  buttons = NULL,
  file_name = "file",
  scrolly = "45vh",
  columnwidth = "200px",
  width = "100\%"
)
}
\arguments{
\item{data}{Dataset.}

\item{caption}{Title for the table.}

\item{digits}{Digits number in the table exported.}

\item{rnames}{Row names.}

\item{buttons}{Buttons: "excel", "copy" or "none". Default c("excel", "copy")}

\item{file_name}{Excel file name}

\item{scrolly}{Windows height to show the table. Default "45vh"}

\item{columnwidth}{Column width. Default '200px'}

\item{width}{Width in pixels or percentage (Defaults to automatic sizing)}
}
\value{
table in markdown format for html documents
}
\description{
Export tables with download, pasta and copy buttons
}
\examples{

\dontrun{

library(inti)

met \%>\%
  web_table(caption = "Web table")

}

}
