% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yupana_import.R
\name{yupana_import}
\alias{yupana_import}
\title{Import information from data summary}
\usage{
yupana_import(data)
}
\arguments{
\item{data}{Summary information with options}
}
\value{
list
}
\description{
Graph summary data
}
\examples{

\dontrun{

library(inti)
library(gsheet)

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "15r7ZwcZZHbEgltlF6gSFvCTFA-CFzVBWwg3mFlRyKPs/edit#gid=1583299871")
# browseURL(url)

fb <- gsheet2tbl(url)

info <- yupana_import(fb)

}

}
