% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yupana_reshape.R
\name{yupana_reshape}
\alias{yupana_reshape}
\title{Fieldbook reshape}
\usage{
yupana_reshape(
  data,
  last_factor,
  sep,
  new_colname,
  from_var = NULL,
  to_var = NULL,
  exc_factors = NULL
)
}
\arguments{
\item{data}{Field book raw data.}

\item{last_factor}{The last factor in your field book.}

\item{sep}{Character that separates the last value.}

\item{new_colname}{The new name for the column created.}

\item{from_var}{The first variable in case you want to exclude several.
variables.}

\item{to_var}{The last variable in case you want to exclude several
variables.}

\item{exc_factors}{Factor to exclude during the reshape.}
}
\value{
data frame
}
\description{
Function to reshape fieldbook according a separation character
}
\details{
If you variable name is \code{variable_evaluation_rep}. The reshape function
will help to create the column \code{rep} and the new variable name will be
\code{variable_evaluation}.
}
