% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_augmented.R
\name{design_augmented}
\alias{design_augmented}
\title{Experimental design: Augmented}
\usage{
design_augmented(
  checks,
  entries,
  blocks = NULL,
  eu_block = NULL,
  random = TRUE,
  zigzag = FALSE,
  dim = NA,
  serie = 1000,
  seed = NULL,
  project = "inkaverse",
  qrcode = "{project}{plots}{entry}"
)
}
\arguments{
\item{checks}{Vector of check treatments.}

\item{entries}{Vector of new entries.}

\item{blocks}{Optional number of blocks. If \code{NULL}, it is calculated from
\code{entries}, \code{checks} and \code{eu_block}.}

\item{eu_block}{Number of experimental units per block.}

\item{random}{Randomize entries allocation.}

\item{zigzag}{Zigzag field layout.}

\item{dim}{Optional layout dimensions c(nrows, ncols).}

\item{serie}{Plot series number.}

\item{seed}{Random seed.}

\item{project}{Barcode prefix.}

\item{qrcode}{QR code template.}
}
\value{
List with fieldbook and parameters.
}
\description{
Fieldbook generator for Augmented Designs.
}
