% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_mvr.R
\name{fieldbook_mvr}
\alias{fieldbook_mvr}
\title{Multivariate Analysis}
\usage{
fieldbook_mvr(data, fb_smr, quali_sup)
}
\arguments{
\item{data}{Field book data.}

\item{fb_smr}{Summary of the variables in the fieldbook.}

\item{quali_sup}{Variables for group the analysis.}
}
\value{
result and plots
}
\description{
Mutilvariate analysis: PCA and HCPC
}
\details{
Compute and plot information for multivarite analysis
}
\examples{

\dontrun{

library(inti)
library(googlesheets4)
library(tidyverse)

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "15r7ZwcZZHbEgltlF6gSFvCTFA-CFzVBWwg3mFlRyKPs/edit#gid=172957346")
# browseURL(url)
gs <- as_sheets_id(url)

(data <- gs \%>\%
    range_read("fb"))

(fb_smr <- gs \%>\%
  range_read("fbsm"))


mvr <- fieldbook_mvr(data, fb_smr, quali_sup = c("genotype", "treat"))

}

}
\author{
Flavio Lozano-Isla
}
