% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook_varlist.R
\name{fieldbook_varlist}
\alias{fieldbook_varlist}
\title{Fieldbook variable list}
\usage{
fieldbook_varlist(fieldbook, varlist = NULL)
}
\arguments{
\item{fieldbook}{Data frame with the fieldbook.}

\item{varlist}{Data frame with the variables information. See examples.}
}
\value{
data frame
}
\description{
Function to include the variables to evaluate in the fieldbook design.
}
\details{
The function allows to include the arguments in the sheet that have
  the information of the variables. You should include 3 columns in the
  sheet: \code{{abbreviation}}, \code{{evaluation}} and \code{{sampling}}.
  See examples. The information will be extracted automatically and deploy
  the list of variable for the fieldbook design. If you will be evaluate the
  experimental unit one time you should set 1.
}
