% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2rmd.R
\name{figure2rmd}
\alias{figure2rmd}
\title{Figure to Rmarkdown}
\usage{
figure2rmd(text, path = ".", opts = NA)
}
\arguments{
\item{text}{Markdown text with figure information (string)}

\item{path}{Image path for figures (path: ".")}

\item{opts}{chunk options in brackets (string: NA)}
}
\value{
string mutated
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
