% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2rmd.R
\name{figure2rmd}
\alias{figure2rmd}
\title{Figure to Rmarkdown format}
\usage{
figure2rmd(text, path = ".", opts = NA)
}
\arguments{
\item{text}{String with the table information}

\item{path}{Path of the image for the figure}

\item{opts}{chunk options in brackets.}
}
\value{
Mutated string
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
