% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{footnotes}
\alias{footnotes}
\title{Table footnote}
\usage{
footnotes(table, notes = NULL, label = "Note:", notation = "alphabet")
}
\arguments{
\item{table}{Kable output in pandoc format.}

\item{notes}{Footnotes for the table.}

\item{label}{Label for start the footnote.}

\item{notation}{Notation for the footnotes (default = "alphabet"). See
details.}
}
\value{
Table with footnotes for word and html documents
}
\description{
Include tables footnotes and symbols for kables in pandoc format
}
\details{
You should use the pandoc format \code{kable(format = "pipe")}. You can add
the footnote symbol using \code{{hypen}} in your table. \code{notation} could
be use: "alphabet", "number", "symbol", "none".
}
