% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc2qmd.R
\name{gdoc2qmd}
\alias{gdoc2qmd}
\title{Google docs to Rmarkdown}
\usage{
gdoc2qmd(file, export = NA, format = "qmd", type = "asis", fill_table = "down")
}
\arguments{
\item{file}{Zip file path from Articul8 exported in md format [path]}

\item{export}{Path to export the files [path: NA (file directory)]}

\item{format}{Output format [string: "qmd" "rmd"]}

\item{type}{output file type [strig: "asis" "list", "listfull", "full"]}

\item{fill_table}{extract table in listfull [strig: "up", "down"]}
}
\value{
path
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
\details{
Document rendering until certain point: "#| end"
Include for next page: "#| newpage"
You can include the cover page params using "#|" in a Google docs table
}
