% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc2qmd.R
\name{gdoc2qmd}
\alias{gdoc2qmd}
\title{Google docs to Rmarkdown}
\usage{
gdoc2qmd(file, export = NA)
}
\arguments{
\item{file}{Zip file path from Articul8 exported in md format [path]}

\item{export}{Path to export the files [path: NA (file directory)]}
}
\value{
path
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
