% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc2qmd.R
\name{gdoc2qmd}
\alias{gdoc2qmd}
\title{Google docs to Rmarkdown}
\usage{
gdoc2qmd(file, export = NA, format = "qmd", type = "asis")
}
\arguments{
\item{file}{Zip file path from Articul8 exported in md format [path]}

\item{export}{Path to export the files [path: NA (file directory)]}

\item{format}{Output format [string: "qmd" "rmd"]}

\item{type}{output file type [strig: "asis" "list", "listfull", "full"]}
}
\value{
path
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
\details{
If you add "|# END" will replace by "knitr::knit_exit()"
}
