% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc2rmd.R
\name{gdoc2rmd}
\alias{gdoc2rmd}
\title{Google docs to Rmarkdown}
\usage{
gdoc2rmd(file, export = "files", md = TRUE)
}
\arguments{
\item{file}{Zip file path from Articul8 exported in  md format}

\item{export}{path to export the files. Default work directory}

\item{md}{Conserve md file. Default = TRUE}
}
\value{
folder
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
