% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdoc2rmd.R
\name{gdoc2rmd}
\alias{gdoc2rmd}
\title{Google docs to Rmarkdown}
\usage{
gdoc2rmd(file, export = "files", prefix_fig = "Figure", prefix_tab = "Table")
}
\arguments{
\item{file}{Zip file path from Articul8 exported in  md format}

\item{export}{path to export the files. Default file directory}

\item{prefix_fig}{Prefix for the name of the figure}

\item{prefix_tab}{Prefix for the name of the table}
}
\value{
folder
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
