% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2cal.R
\name{H2cal}
\alias{H2cal}
\title{Heritability in plant breeding on a genotype-difference basis}
\source{
https://github.com/PaulSchmidtGit/Heritability/tree/master/Alternative%20Heritability%20Measures

https://stackoverflow.com/questions/38697477/mean-variance-of-a-difference-of-blues-or-blups-in-lme4
}
\usage{
H2cal(
  data,
  trait,
  gen.name,
  rep.n,
  loc.n = 1,
  year.n = 1,
  loc.name = NULL,
  year.name = NULL,
  fix.model,
  ran.model,
  summary = FALSE,
  blues = FALSE,
  effects = FALSE,
  plot_diag = FALSE,
  plot_dots = NULL,
  outliers.rm = FALSE
)
}
\arguments{
\item{data}{Experimental design data frame with the factors and traits.}

\item{trait}{Name of the trait.}

\item{gen.name}{Name of the genotypes.}

\item{rep.n}{Number of replications in the experiment.}

\item{loc.n}{Number of locations (default = 1). See details.}

\item{year.n}{Number of years (default = 1). See details.}

\item{loc.name}{Name of the location (default = NULL). See details.}

\item{year.name}{Name of the years (default = NULL). See details.}

\item{fix.model}{The fixed effects in the model. See examples.}

\item{ran.model}{The random effects in the model. See examples.}

\item{summary}{Summary of the random model (default = TRUE).}

\item{blues}{Calculate the BLUEs (default = FALSE).}

\item{effects}{Conditional modes of the random effects instead of the BLUPs (default = FALSE).}

\item{plot_diag}{Show diagnostic plots (default = FALSE).}

\item{plot_dots}{Show dotplot genotypes vs trait (default = NULL). See examples.}

\item{outliers.rm}{Remove outliers (default = FALSE)}
}
\value{
A list with three object:

1. Table with the variance components and heritability.

2. BLUPs.

3. BLUEs.
}
\description{
Function to calculate:
1. The variance components.
2. Heritability under three approaches: Standard, Cullis and Piepho.
3. Best Linear Unbiased Predictors (BLUPs).
4. Best Linear Unbiased Estimators (BLUEs).
}
\details{
The function allows to made the calculation for individual or multi-environmental trials (MET) using th fixed and random model.

For individual experiments is necessary provide the \code{trait}, \code{gen.name}, \code{rep.n}.

For MET experiments you should \code{loc.n} and \code{loc.name} y/o \code{year.n} and \code{year.name} according your experiment.

The blues calculation is based in the pairwise comparison and its could takes time according the number of the genotypes.

You can specify as \code{blues = FALSE} for calculate the variance components and blups faster.

For more information review the references.
}
\examples{

\dontrun{
library(inti)
library(tidyverse)
library(emmeans)
library(lme4)
library(lmerTest)
library(agridat)

 dt <- john.alpha
 hr <- H2cal(data = dt
            , trait = "yield"
            , gen.name = "gen"
            , rep.n = 3
            , fix.model = "rep + (1|rep:block) + gen"
            , ran.model = "rep + (1|rep:block) + (1|gen)"
            , blues = T
            , plot_diag = TRUE
            , plot_dots = "rep"
            )
 hr$tabsmr
 hr$blups
 hr$blues
}

}
\references{
Schmidt, P., J. Hartung, J. Bennewitz, and H.-P. Piepho. 2019. Heritability in Plant Breeding on a Genotype-Difference Basis. Genetics 212(4): 991–1008. doi: 10.1534/genetics.119.302134.

Schmidt, P., J. Hartung, J. Rath, and H.-P. Piepho. 2019. Estimating Broad-Sense Heritability with Unbalanced Data from Agricultural Cultivar Trials. Crop Science 59(2): 525–536. doi: 10.2135/cropsci2018.06.0376.

Bernal-Vasquez, Angela-Maria, et al. “Outlier Detection Methods for Generalized Lattices: A Case Study on the Transition from ANOVA to REML.” Theoretical and Applied Genetics, vol. 129, no. 4, Apr. 2016, pp. 787–804. Springer Link, doi:10.1007/s00122-016-2666-6.
}
\author{
Maria Belen Kistner

Flavio Lozano-Isla
}
