% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_figure.R
\name{include_figure}
\alias{include_figure}
\title{Figure with caption and notes}
\usage{
include_figure(
  data,
  figure,
  fig_list = "figure",
  path = "url",
  caption = "caption",
  notes = "notes",
  label = "__Source:__"
)
}
\arguments{
\item{data}{data frame with the figures information. See details.}

\item{figure}{name of the figure in the fig_list.}

\item{fig_list}{column with the list of figures (default = "figure").}

\item{path}{figure path or url (default = "url").}

\item{caption}{column with the figure caption (default = "caption").}

\item{notes}{column with the figure notes (default = "notes").}

\item{label}{label for start the footnote (default = "__Source:__").}
}
\value{
Figure with caption and notes
}
\description{
Include figures with title and notes using a data base
}
\details{
The data frame should contain 4 columns:
1. \code{fig_list}
2. \code{caption}
3. \code{notes}
4. \code{path} or \code{url}
}
\author{
Flavio Lozano-Isla
}
