% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_figure.R
\name{include_figure}
\alias{include_figure}
\title{Figure with caption and notes}
\usage{
include_figure(data = NULL, figure, caption = NA, notes = NA, label = "Note:")
}
\arguments{
\item{data}{Data frame with the figures information. See details.}

\item{figure}{Name or path of the figure.}

\item{caption}{Manual figure caption (default = NA).}

\item{notes}{Manual figure notes (default = NA).}

\item{label}{Label for start the footnote (default = "Note:").}
}
\value{
Figure with caption and notes
}
\description{
Include figures with title and notes using a data base
}
\details{
The data frame should contain 4 rows in \code{table}:

1. \code{caption}.

2. \code{notes}.

3. \code{url} 

4. \code{path}.

If you don't use a data frame you provide the information manually.
}
\examples{

library(googlesheets4)
library(inti)

finfo <- info_figure(caption = "test"
                   , notes = "nota"
                   , url = "https://devblackops.io/images/testing.jpg"
                   , path = "test.jpg"
                   )

fig <- finfo \%>\%  include_figure()
fig


}
