% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_figure.R
\name{include_figure}
\alias{include_figure}
\title{Figure with caption and notes}
\usage{
include_figure(
  data = NULL,
  figure,
  caption = NULL,
  notes = NULL,
  label = "Source:",
  table = c("figures", "caption", "notes", "url")
)
}
\arguments{
\item{data}{data frame with the figures information. See details.}

\item{figure}{name or path of the figure.}

\item{caption}{manual figure caption (default = NULL).}

\item{notes}{manual figure notes (default = NULL).}

\item{label}{label for start the footnote (default = "Source:").}

\item{table}{columns in the table (default = c("figures", "caption", "notes", "url")).}
}
\value{
Figure with caption and notes
}
\description{
Include figures with title and notes using a data base
}
\details{
The data frame should contain 4 columns specified in \code{table}:
1. \code{figures}.
2. \code{caption}.
3. \code{notes}.
4. \code{path} or \code{url}.
If you don't use a a data frame you provide the information manually.
}
\author{
Flavio Lozano-Isla
}
