% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_table.r
\name{include_table}
\alias{include_table}
\title{Table with footnotes}
\usage{
include_table(
  data,
  caption = NULL,
  notes = NULL,
  label = "__Note:__",
  notation = "number"
)
}
\arguments{
\item{data}{data frame.}

\item{caption}{table caption. See details.}

\item{notes}{footnotes for the table. See details.}

\item{label}{label for start the footnote (default = "__Note:__").}

\item{notation}{notation for the symbols and footnotes (default = "number"). See details.}
}
\value{
Table with caption and footnotes
}
\description{
Include tables with notes and titles for word and html documents
}
\details{
For \code{caption} and \code{notes} you can include [caption] and [notes] in the last rows of your data frame.
Where the information will be extracted automatically for include in the formatted table.
You can add the footnote symbol using \code{[fn]} in your table.
\code{notation} could be use: "alphabet", "number", "symbol", "none".
}
\author{
Flavio Lozano-Isla
}
