% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jc_tombola.R
\name{jc_tombola}
\alias{jc_tombola}
\title{Journal Club Tombola}
\usage{
jc_tombola(
  data,
  members,
  papers = 1,
  group = NA,
  gr_lvl = NA,
  status = NA,
  st_lvl = "active",
  frq = 7,
  date = NA,
  seed = NA
)
}
\arguments{
\item{data}{Data frame withe members and their information.}

\item{members}{Columns with the members names.}

\item{papers}{Number of paper by meeting}

\item{group}{Column for arrange the group.}

\item{gr_lvl}{Levels in the groups for the arrange. See details.}

\item{status}{Column with the status of the members.}

\item{st_lvl}{Level to confirm the assistance in the JC. See details.}

\item{frq}{Number of the day for each session.}

\item{date}{Date when start the first session of JC.}

\item{seed}{Number for replicate the results (default = date).}
}
\value{
data frame with the schedule for the JC
}
\description{
Function for arrange journal club schedule
}
\details{
The function could consider n levels for \code{gr_lvl}. In the case of more
levels using "both" or "all" will be the combination. The suggested levels 
for \code{st_lvl} are: active or spectator. Only the "active" members will 
enter in the schedule.
}
