% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Regresion line equation}
\usage{
lm_eqn(x, y, data)
}
\arguments{
\item{x}{variable in the x axis}

\item{y}{variable in the y axis}

\item{data}{dataframe with the information}
}
\value{
regression equation
}
\description{
Construc the regression line equation
}

