% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_comparison.R
\name{mean_comparison}
\alias{mean_comparison}
\title{Mean comparison test}
\usage{
mean_comparison(
  data,
  fb_smr,
  variable,
  model_facts,
  comp_facts,
  test_comp = "SNK",
  sig_level = 0.05,
  graph_opts = FALSE
)
}
\arguments{
\item{data}{Field book data.}

\item{fb_smr}{Summary of the variables in the fieldbook.}

\item{variable}{Model used for the experimental design.}

\item{model_facts}{Comparison test (default = "SNK"). Others: "TUKEY" & "DUNCAN".}

\item{comp_facts}{Significance level for the analysis (default = 0.05).}

\item{test_comp}{Significance level for the analysis (default = 0.05).}

\item{sig_level}{Significance level for the analysis (default = 0.05).}

\item{graph_opts}{Include option in the table for graphs (default = FALSE)}
}
\value{
data frame
}
\description{
Function to compare treatment from lm or aov
}
\details{
For compare the factors you should use ":". For example, to compare treatment1
and treatment2 => "treatment1:treatment2".
}
\examples{

\dontrun{

library(inti)
library(googlesheets4)
library(tidyverse)

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "15r7ZwcZZHbEgltlF6gSFvCTFA-CFzVBWwg3mFlRyKPs/edit#gid=1414357945")
# browseURL(url)
gs <- as_sheets_id(url)

(data <- gs \%>\%
    range_read("fb"))

(fb_smr <- gs \%>\%
  range_read("fbsm"))

mc <- mean_comparison(data
                     , fb_smr = fb_smr
                     , variable = "BIOMDW"
                     , graph_opts = T
                     )

table <- mc$comparison

table \%>\% sheet_write(ss = gs, sheet = "plot")

}

}
\author{
Flavio Lozano-Isla
}
