% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamorfosis.R
\name{metamorphosis}
\alias{metamorphosis}
\title{Transform data frame based in a dictionary}
\usage{
metamorphosis(fieldbook, dictionary, from, to, index, colnames)
}
\arguments{
\item{fieldbook}{Data frame with the original information.}

\item{dictionary}{Data frame with new names and categories. See details.}

\item{from}{Column of the dictionary with the original names.}

\item{to}{Column of the dictionary with the new names.}

\item{index}{Column of the dictionary with the type and level of the
variables.}

\item{colnames}{Character vector with the name of the columns.}
}
\value{
List with two objects. 1. New data frame. 2. Dictionary.
}
\description{
Transform entire fieldbook according to data a dictionary
}
\details{
The function require at least three columns.

  1. Original names (\code{from}).

  2. New names (\code{to}).

  3. Variable type (\code{index}).
}
