% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\title{Fieldbook plot experimental designs}
\usage{
plot_design(
  data,
  factors,
  dim = NULL,
  fill = "plots",
  colour = NULL,
  xlab = NULL,
  ylab = NULL,
  glab = NULL
)
}
\arguments{
\item{data}{Experimental design data frame with the factors and level. See examples.}

\item{factors}{Vector with the name of the columns with the factors.}

\item{dim}{Dimension for reshape the design arrangement.}

\item{fill}{Value for fill the experimental units (default = "plots").}

\item{colour}{Color for the experimental units (default = first factor).}

\item{xlab}{Title for x axis}

\item{ylab}{Title for y axis}

\item{glab}{Title for group axis}
}
\value{
plot
}
\description{
Function to plot the fieldbook design
}
\details{
The function allows to plot the experimental design according the
field experiment design.
}
\examples{

\dontrun{

library(inti)
library(googlesheets4)
library(tidyverse)

url <- paste0("https://docs.google.com/spreadsheets/d/"
, "1ilw0NHT7mihaM-3U48KzkuMt927xe8ukX6rNuIw2fT0/edit#gid=0")
# browseURL(url)
gs <- as_sheets_id(url)

(fb <- gs \%>\%
    range_read("tarpuy"))

data <- fb \%>\% inti::fieldbook_design(n_factors = 3
                                      , type = "rcbd"
                                      , rep = 3) \%>\% pluck("design")
plot_design(data
            , factors = c("temp", "prof", "condición")
            , dim = "block"
            , glab = "Temperature"
            )

}

}
\author{
Flavio Lozano-Isla
}
