% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smr.R
\name{plot_smr}
\alias{plot_smr}
\title{Plot summary data}
\usage{
plot_smr(
  data,
  type = c("bar", "line"),
  x,
  y,
  groups,
  xlab,
  ylab,
  glab,
  limits,
  brakes,
  sig,
  error,
  legend = "top"
)
}
\arguments{
\item{data}{data frame with summary data. See details}

\item{type}{type of plot  (default = "bar"). Others: "line".}

\item{x}{Variable in the x axis.}

\item{y}{Variable in the y axis.}

\item{groups}{Groups the x axis (legend) .}

\item{xlab}{Label of x axis.}

\item{ylab}{Label of y axis.}

\item{glab}{Label of groups or legend.}

\item{limits}{Limits max and min.}

\item{brakes}{Units for break the limits.}

\item{sig}{Comparison test values (default = "sig").}

\item{error}{Variable used for error bars .}

\item{legend}{Position of legend (default = "top"). Others: "left", "right", "bottom", "none"}
}
\value{
plot
}
\description{
Graph summary data into bar o line plot
}
\details{
If the table is a out put of `mean_comparison(graph_opts = TRUE)` function.
Its contain all the parameter for the plot
}
\examples{

\dontrun{

library(inti)
library(googlesheets4)
library(tidyverse)

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "15uwCgQRtR01B3FJaZBE8t_0bOC_8Bbey9ccwVlZH0jg/edit#gid=56711214")
# browseURL(url)
gs <- as_sheets_id(url)

(data <- gs \%>\%
    range_read("plot"))

plot_smr(data, type = "bar")

}

}
\author{
Flavio Lozano-Isla
}
