% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_split_rcbd_design.R
\name{plot_split_rcbd_design}
\alias{plot_split_rcbd_design}
\title{Plot split-plot RCBD fieldbook design}
\usage{
plot_split_rcbd_design(
  data,
  factor = NA,
  fill = "plots",
  xlab = NULL,
  ylab = NULL,
  glab = NULL
)
}
\arguments{
\item{data}{Fieldbook data frame from \code{design_split_rcbd()}.}

\item{factor}{Character. Column used to color experimental units.}

\item{fill}{Character vector. Column or columns used as labels inside each
experimental unit.}

\item{xlab}{Character. Optional x axis title.}

\item{ylab}{Character. Optional y axis title.}

\item{glab}{Character. Optional legend title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot fieldbook sketches for split-plot designs under RCBD structure.
}
