% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rticles.R
\name{rticles}
\alias{rticles}
\title{Build markdown o bookdown template}
\usage{
rticles()
}
\value{
Shiny app
}
\description{
Invoke RStudio add-in to create markdown documents
}
\details{
Create all the files in the present work directory.
It is recommended use ".Rproj" for bookdown documents.
After create the ".Rproj" file.
Open the ".Rprj" and you should have the possibility to compile the book.
}
\examples{
\dontrun{
 rticles()
}

}
\author{
Flavio Lozano-Isla
}
