% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2qmd.R
\name{table2qmd}
\alias{table2qmd}
\title{Table to Rmarkdown}
\usage{
table2qmd(text)
}
\arguments{
\item{text}{Markdown text with table information (string)}
}
\value{
string mutated
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
