% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2rmd.R
\name{table2rmd}
\alias{table2rmd}
\title{Table to Rmarkdown}
\usage{
table2rmd(text)
}
\arguments{
\item{text}{Markdown text with table information (string)}
}
\value{
string mutated
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
