% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2rmd.R
\name{table2rmd}
\alias{table2rmd}
\title{Table to Rmarkdown}
\usage{
table2rmd(text, opts = NA, prefix = "Table")
}
\arguments{
\item{text}{String with the table information}

\item{opts}{chunk options in brackets.}

\item{prefix}{Prefix for the name of the table}
}
\value{
Mutated string
}
\description{
Use Articul8 Add-ons from Google docs to build Rticles
}
