% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarpuy.R
\name{tarpuy}
\alias{tarpuy}
\title{Field book experimental designs}
\usage{
tarpuy()
}
\value{
Shiny app
}
\description{
Invoke RStudio addin to create field book designs
}
\details{
Tarpuy allow to create experimental designs under an interactive app.
}
\examples{
\dontrun{

 inti::tarpuy()

}

}
\author{
Flavio Lozano-Isla
}
