% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarpuy.R
\name{tarpuy}
\alias{tarpuy}
\title{Fieldbook experimental designs}
\usage{
tarpuy(dependencies = FALSE)
}
\arguments{
\item{dependencies}{Install package dependencies for run the app}
}
\value{
Shiny app
}
\description{
Invoke RStudio addin to create field book designs
}
\details{
Tarpuy allow to create experimental designs under an interactive app.
}
\examples{

if(interactive()){

 inti::tarpuy()

}

}
