% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarpuy_plex.R
\name{tarpuy_plex}
\alias{tarpuy_plex}
\title{Fieldbook plan information}
\usage{
tarpuy_plex(
  data = NULL,
  title = NULL,
  short_title = NULL,
  objective = NULL,
  references = NULL,
  plan = NULL,
  institutions = NULL,
  researchers = NULL,
  manager = NULL,
  location = NULL,
  altitude = NULL,
  georeferencing = NULL,
  environment = NULL,
  start = NA,
  end = NA,
  project = NULL,
  repository = NULL,
  manuscript = NULL,
  album = NULL,
  nfactor = 2,
  design = "rcbd",
  rep = 4,
  zigzag = FALSE,
  nrows = NA,
  serie = 1000,
  seed = 0,
  qrcode = "{project}{plots}",
  aug_blocks = NA,
  aug_eu_block = NA,
  aug_random = TRUE
)
}
\arguments{
\item{data}{Data with the fieldbook information.}

\item{title}{Project title.}

\item{short_title}{Short description of the project.}

\item{objective}{The objectives of the project.}

\item{references}{References.}

\item{plan}{General description of the project (M & M).}

\item{institutions}{Institutions involved in the project.}

\item{researchers}{Persons involved in the project.}

\item{manager}{Persons responsible of the collection of the data.}

\item{location}{Location of the project.}

\item{altitude}{Altitude of the experiment (m.a.s.l).}

\item{georeferencing}{Georeferencing information.}

\item{environment}{Environment of the experiment (greenhouse, lab, etc).}

\item{start}{The date of the start of the experiments.}

\item{end}{The date of the end of the experiments.}

\item{project}{Name or ID for the fieldbook/project.}

\item{repository}{link to the repository.}

\item{manuscript}{link for manuscript.}

\item{album}{link with the photos of the project.}

\item{nfactor}{Number of factors for the design.}

\item{design}{Type of design.}

\item{rep}{Number of replication.}

\item{zigzag}{Experiment layout in zigzag \verb{[logic: FALSE]}}

\item{nrows}{Experimental design dimension by rows \verb{[numeric: value]}}

\item{serie}{Number of digits in the plots.}

\item{seed}{Seed for the randomization. A value of \code{0}, \code{NA}, or \code{NULL}
generates and stores an effective seed in the design sheet.}

\item{qrcode}{QR code template used to concatenate fieldbook identifiers.}

\item{aug_blocks}{Optional number of blocks for augmented design.}

\item{aug_eu_block}{Number of plots per block for augmented design.}

\item{aug_random}{Logical. Randomize entries allocation in augmented design.}
}
\value{
data frame or list of arguments:

\enumerate{ \item info \item variables \item design \item logbook \item
timetable \item budget }
}
\description{
Information for build a plan for an experiment (PLEX)
}
\details{
Provide the information available.
}
