% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarpuy_plotdesign.R
\name{tarpuy_plotdesign}
\alias{tarpuy_plotdesign}
\title{Fieldbook plot experimental designs}
\usage{
tarpuy_plotdesign(
  data,
  factor = NA,
  fill = "plots",
  xlab = NULL,
  ylab = NULL,
  glab = NULL
)
}
\arguments{
\item{data}{Fieldbook data frame or design object containing a fieldbook.}

\item{factor}{Character. Column used to color experimental units.}

\item{fill}{Character vector. Column or columns used as labels inside
experimental units.}

\item{xlab}{Character. Optional x axis title.}

\item{ylab}{Character. Optional y axis title.}

\item{glab}{Character. Optional legend title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot fieldbook sketches according to the experimental design type.
}
\details{
This function works as a dispatcher. It detects the design type from the
fieldbook and sends the data to the corresponding plotting function.
}
