% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarpuy_plotdesign.R
\name{tarpuy_plotdesign}
\alias{tarpuy_plotdesign}
\title{Fieldbook plot experimental designs}
\usage{
tarpuy_plotdesign(
  data,
  factor,
  dim = NULL,
  fill = "plots",
  xlab = NULL,
  ylab = NULL,
  glab = NULL
)
}
\arguments{
\item{data}{Experimental design data frame with the factors and level. See
examples.}

\item{factor}{Vector with the name of the columns with the factors.}

\item{dim}{Dimension for reshape the design arrangement.}

\item{fill}{Value for fill the experimental units (default = "plots").}

\item{xlab}{Title for x axis.}

\item{ylab}{Title for y axis.}

\item{glab}{Title for group axis.}
}
\value{
plot
}
\description{
Plot fieldbook sketch designs based in experimental design
}
\details{
The function allows to plot the experimental design according the
  field experiment design.
}
