% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_table.R
\name{web_table}
\alias{web_table}
\title{HTML tables for markdown documents}
\usage{
web_table(data, caption = NULL, digits = 2, rnames = FALSE, buttons = NULL)
}
\arguments{
\item{data}{Dataset.}

\item{caption}{Title for the table.}

\item{digits}{Digits number in the table exported.}

\item{rnames}{Row names.}

\item{buttons}{Buttons: "excel", "copy" or "none". Default c("excel", "copy")}
}
\value{
table in markdown format for html documents
}
\description{
Export tables with download, pasta and copy buttons
}
