% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yupana.R
\name{yupana}
\alias{yupana}
\title{Interactive data analysis}
\usage{
yupana()
}
\value{
Shiny app
}
\description{
Invoke RStudio addin to analyze and graph experimental design data
}
\details{
Yupana: data analysis and graphics for experimental designs.
}
\examples{

if(interactive()){

 inti::yupana()

}

}
